requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
dj-places==4.0.0
Django==3.2.5
django-allauth==0.45.0
django-common-helpers==0.9.2
django-cron==0.5.1
django-crontab==0.7.1
django-materialize-css==0.0.1
django-materializecss-form==1.1.17
django-phonenumber-field==5.2.0
future==0.18.2
geocoder==1.38.1
idna==3.2
kombu==5.1.0
oauthlib==3.1.1
phonenumbers==8.12.29
Pillow==8.3.1
prompt-toolkit==3.0.20
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
ratelim==0.1.6
redis==3.5.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
